+++ proposal.go
@@ -70,7 +70,7 @@ type BaseProposal struct {
 	OverridePreviousRoot bool                                        `json:"overridePreviousRoot"`
 	ChainMetadata        map[types.ChainSelector]types.ChainMetadata `json:"chainMetadata" validate:"required,min=1"`
 	Description          string                                      `json:"description"`
-
+	Metadata             map[string]any                              `json:"metadata,omitempty"`
 	// This field is passed to SDK implementations to indicate whether the proposal is being run
 	// against a simulated environment. This is only used for testing purposes.
 	useSimulatedBackend bool `json:"-"`
@@ -390,4 +390,22 @@ func proposalValidateBasic(proposalObj Proposal) error {
 // merkle tree generation.
 func wrapTreeGenErr(err error) error {
 	return fmt.Errorf("merkle tree generation error: %w", err)
+}
+
+func mergeMetadata(m1, m2 map[string]any) (map[string]any, error) {
+	if len(m2) == 0 {
+		return m1, nil
+	}
+
+	merged := make(map[string]any, len(m1))
+	maps.Copy(merged, m1)
+	for k, v2 := range m2 {
+		v1, exists := m1[k]
+		if exists && v1 != v2 {
+			return nil, fmt.Errorf("conflicting metadata for key %s: %v vs %v (%T %T %v)", k, v1, v2, v1, v2, v1 == v2)
+		}
+		merged[k] = v2
+	}
+
+	return merged, nil
 }