+++ sdk/solana/timelock_converter.go
@@ -13,7 +13,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/gagliardetto/solana-go"
-	bindings "github.com/smartcontractkit/chainlink-ccip/chains/solana/gobindings/timelock"
+	bindings "github.com/smartcontractkit/chainlink-ccip/chains/solana/gobindings/v0_1_1/timelock"
 
 	"github.com/smartcontractkit/mcms/sdk"
 	"github.com/smartcontractkit/mcms/types"
@@ -22,6 +22,11 @@ import (
 var _ sdk.TimelockConverter = (*TimelockConverter)(nil)
 
 type TimelockConverter struct{}
+
+// NewTimelockConverter creates a new instance of the TimelockConverter
+func NewTimelockConverter() *TimelockConverter {
+	return &TimelockConverter{}
+}
 
 func (t TimelockConverter) ConvertBatchToChainOperations(
 	ctx context.Context,